# swagger.json
{
    "swagger":"2.0",
    "info":{
        "description":"A node project to example",
        "version":"1.0.0",
        "title":"Notes"
    },
    "basePath":"",
...
    "schemes":[
        "http"
    ],
    "paths":{
...
        "/note":{
            "post":{
                "tags":[
                    "note"
                ],
                "security": {
                    "bearerAuth": []
                },
                "summary":"Create a new note in system",
                "description":"",
                "operationId":"addNote",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"Note Object",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/Note"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"successful operation",
                        "schema":{
                            "type":"object",
                            "$ref":"#/definitions/Response"                          
                        }
                    },
                    "400":{
                        "description":"Invalid param",
                        "schema":{
                            "type":"object",
                            "$ref":"#/definitions/Response"                          
                        }
                    },
                    "401":{
                        "description":"Invalid token",
                        "schema":{
                            "type":"object",
                            "$ref":"#/definitions/Response"                          
                        }
                    },
                    "500":{
                        "description":"Server Error",
                        "schema":{
                            "type":"object",
                            "$ref":"#/definitions/Response"                          
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth":{
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    },
    "definitions":{
        "Response":{
            "type":"object",
            "properties":{
                "statusCode":{
                    "type":"string"
                },
                "data":{
                    "type":"object"
                }
            }
        },
        "Note":{
            "type":"object",
            "properties":{
                "name":{
                    "type":"string"
                },
                "description":{
                    "type":"string"
                },
                "idUser":{
                    "type":"string"
                }
            }
        },
...
    }
}